Update style.css
Added CSS to include categories on the home page (index.html). These include creating and scaling the boxes appropriately, adding a link when the box is clicked, adding placeholder images for later use, setting the box color to a greyish-white, and aligning text at the bottom of the box for the category name.