Added categories
Added 4 category placeholders for later use. These 4 categories are organized as rectangles and are contained within a link, so once clicked they can link to another webpage. Each category has a placeholder image that is centered, as well as text centered below the image. All of the categories are contained in one <div> in order to format the boxes as well as everything that will go below it.