Added generic product page
Added a generic product page that could be used for any product. This file includes the same header and JavaScript from index.html. A <div> element was created that any text can be added about the product name and some selling points, and a product image was added to the right of the text. Boxes were added to provide a link any retail store from where the product can be purchased from, background color currently set to gray.