Update style.css
All added elements added are for the productpage.html. Added CSS elements create a container for the product description, sets the margin for the text, sets the image to float to the right of the text, and adds the boxes that will contain links to the retail stores that sell each given product.